refactor and simplify xmlrpc error notices
